Adopt uint8 ParkEntranceSequence centre/left/right values 0/1/2 across element accessors, object image selection, importers, placement, removal and rendering. Preserve low-nibble storage and replace the placement loop with the same ordered three values. Adapt fork topology bounds and both test fixtures; expand the real-plugin test to 15 sequence inputs.

Disposition: adopt applicable changes
Upstream-Commit: 7b1b014fce773472ce2eef594bd8df7ef3ba9979
Upstream-Remaining: 187
Migration-Log: docs/upstream-migration-2026-09-13.md

+++ src/openrct2/world/MapPathTopology.cpp
@@ -149,7 +149,7 @@ namespace OpenRCT2::MapPathTopology
         {
             const auto entranceType = entrance.getEntranceType();
             const auto sequence = entrance.getSequenceIndex();
-            if (entranceType > EntranceType::parkEntrance || sequence >= 8)
+            if (entranceType > EntranceType::parkEntrance || static_cast<uint8_t>(sequence) >= 8)
             {
                 isExact = false;
                 return 0;
